Improve error handling and add progress indication
Modified `showTransferMenu` in `cli.h` to return 5 instead of throwing a runtime error when the state is not `CLIState::SESSION`. Added a progress bar manager call in `UploadDirectory` function of `file_transfer_client.cpp` to indicate the start of scanning a directory. Wrapped `UploadFile` call in a `try` block to continue with the next file if an exception is caught, instead of stopping the upload process.